+++ src/components/button/Button.js
@@ -17,7 +17,7 @@ export class ButtonComponent extends BaseComponent {
   set loading(loading) {
     this._loading = loading;
     if (!this.loader && loading) {
-      this.loader = this.ce('buttonLoader', 'i', {
+      this.loader = this.ce('i', {
         class: 'glyphicon glyphicon-refresh glyphicon-spin button-icon-right'
       });
     }
@@ -37,7 +37,7 @@ export class ButtonComponent extends BaseComponent {
   }
 
   build() {
-    this.element = this.ce('element', this.info.type, this.info.attr);
+    this.element = this.ce(this.info.type, this.info.attr);
     if (this.component.label) {
       this.label = this.text(this.component.label);
       this.element.appendChild(this.label);