+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+### Changed
+ - Replace Handlebars templating with lodash.
+
 ## 2.11.8
 ### Fixed
  - Issue where choices.js changed internal api that we rely on.